# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000" #Api runs locally in port 3000
info:
  title: WASAPhoto
  description: "WASAPhoto API documentation"
  version: 1.0.0
components:
  securitySchemes:  #Security of API, for authentication use Bearer where token represented in JWT Format
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /group/name: #6)giving a name group
    post:
      tags:
        - Groups
      summary: "Set group name"
      description: "Allows an admin to update the name of a group."
      operationId: setGroupName
      security:
        - BearerAuth: []
      requestBody:
        description: "New group name"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request body for updating the group's name"
              properties:
                groupId:
                  type: integer
                  format: int64
                  description: "ID of the group"
                  example: 123
                groupName:
                  type: string
                  description: "The new group name"
                  example: "New Group Name"
                  minLength: 1
                  maxLength: 100
                  pattern: "^[\\w\\s]+$"
              required:
                - groupId
                - groupName
      responses:
        '200':
          description: "Group name updated successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating the group name was updated"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the operation was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid input)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /group/leave: #9)leaving a group
    post:
      tags:
        - Groups
      summary: "Leave a group"
      description: "Allows a user to leave a specified group."
      operationId: leaveGroup
      security:
        - BearerAuth: []
      requestBody:
        description: "Group leave request"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for leaving a group"
              properties:
                groupId:
                  type: integer
                  format: int64
                  description: "ID of the group to leave"
                  example: 101
              required:
                - groupId
      responses:
        '200':
          description: "Successfully left the group"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating successful group leave"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the operation was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid request body)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid JSON format"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
  /message/send:
    post:
      tags:
        - Messages
      summary: "Send a message"
      description: "Allows a user to send a message to another user or a group."
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Message sending request"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for sending a message"
              properties:
                recipientId:
                  type: integer
                  format: int64
                  description: "ID of the recipient (user or group)"
                  example: 456
                isGroup:
                  type: boolean
                  description: "Specifies whether the recipient is a group"
                  example: false
                content:
                  type: string
                  description: "The message content"
                  example: "Hello, how are you?"
                  minLength: 1
                  maxLength: 1000
                  pattern: "^[\\s\\S]+$"
              required:
                - recipientId
                - isGroup
                - content
      responses:
        '200':
          description: "Message sent successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing details of the sent message"
                properties:
                  messageId:
                    type: integer
                    format: int64
                    description: "ID of the sent message"
                    example: 789
                  senderId:
                    type: integer
                    format: int64
                    description: "ID of the sender"
                    example: 123
                  recipientId:
                    type: integer
                    format: int64
                    description: "ID of the recipient"
                    example: 456
                  content:
                    type: string
                    description: "The message content"
                    example: "Hello, how are you?"
                    minLength: 1
                    maxLength: 1000
                    pattern: "^[\\s\\S]+$"
                  timestamp:
                    type: string
                    format: date-time
                    description: "Timestamp of when the message was sent"
                    example: "2025-02-11T15:00:00Z"
                    minLength: 20
                    maxLength: 30
                    pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
                required:
                  - messageId
                  - senderId
                  - recipientId
                  - content
                  - timestamp
        '400':
          description: "Bad request (invalid request payload)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Failed to send message"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message/read:
    post:
      tags:
        - Messages
      summary: "Mark messages as read"
      description: "Marks all messages in a conversation as read by the user."
      operationId: markAsRead
      security:
        - BearerAuth: []
      requestBody:
        description: "Request to mark messages as read"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for marking messages as read"
              properties:
                conversationId:
                  type: integer
                  format: int64
                  description: "ID of the conversation"
                  example: 123
              required:
                - conversationId
      responses:
        '200':
          description: "Messages marked as read successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating messages were marked as read"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the operation was successful"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (failed to mark as read)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Failed to mark messages as read"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
  /message/send-photo:
    post:
      tags:
        - Messages
      summary: "Upload a profile photo"
      description: "Allows a user to upload a new profile photo."
      operationId: sendPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: "Upload profile photo"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Object containing the profile photo file"
              properties:
                profilePicture:
                  type: string
                  format: binary
                  description: "The profile photo file to upload"
                  minLength: 1
                  maxLength: 256
              required:
                - profilePicture
      responses:
        '200':
          description: "Profile photo uploaded successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing the uploaded photo path"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the upload was successful"
                    example: true
                  photoPath:
                    type: string
                    description: "URL or path of the uploaded photo"
                    example: "/uploads/user123/profile.jpg"
                    minLength: 5
                    maxLength: 256
                    pattern: "^/uploads/.+$"
                required:
                  - success
                  - photoPath
        '400':
          description: "Bad request (failed to upload photo)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Failed to upload profile picture"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...